fix: change mgsbacklog type

+++ pulsar-admin-jdk/src/main/java/io/github/protocol/pulsar/admin/jdk/SubscriptionStats.java
@@ -32,7 +32,7 @@ public class SubscriptionStats {
 
     public int chunkedMessageRate;
 
-    public int msgBacklog;
+    public long msgBacklog;
 
     public long backlogSize;
